requirements.txt
python-telegram-bot==20.8
feedparser==6.0.10
requests==2.31.0
python-dateutil==2.8.2